refactor(http11): 디렉토리 분리
